pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "thess-geo-analytics"
version = "0.2.0"
description = "Reproducible EO + geospatial analytics pipeline for Thessaloniki (EL522)."
requires-python = ">=3.11"

dependencies = [
  # Core utilities
  "python-dotenv>=1.0.0",
  "requests>=2.31.0",
  "numpy>=1.24",
  "pandas>=2.0",
  "pyarrow>=14.0.1",

  # Geospatial stack
  "geopandas>=0.14",
  "shapely>=2.0",
  "pyproj>=3.6",
  "pyogrio>=0.7.2",

  # STAC / EO catalog access
  "pystac-client>=0.7.6",
  "pystac>=1.10.0",

  # Raster / array processing
  "rasterio>=1.3.10",
  "xarray>=2024.1.0",
  "netCDF4>=1.6",
  "dask[array]>=2024.1.0",
  "rioxarray>=0.15.0",

  # Visualization / progress
  "matplotlib>=3.8",
  "tqdm>=4.66",

  "pyyaml>=6.0",              
  "scipy>=1.10",              
  "scikit-image>=0.22",        
  "google-cloud-storage>=2.0", 
  "sentinelhub>=3.11.0",       
  "utm>=0.8.1",               
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]